add lazy loading capability and u64 position

+++ src/iter.rs
@@ -5,18 +5,18 @@ pub struct LineColIterator<I> {
 
     /// Index of the current line. Characters in the first line of the input
     /// (before the first newline character) are in line 1.
-    line: usize,
+    line: u64,
 
     /// Index of the current column. The first character in the input and any
     /// characters immediately following a newline character are in column 1.
     /// The column is 0 immediately after a newline character has been read.
-    col: usize,
+    col: u64,
 
     /// Byte offset of the start of the current line. This is the sum of lengths
     /// of all previous lines. Keeping track of things this way allows efficient
     /// computation of the current line, column, and byte offset while only
     /// updating one of the counters in `next()` in the common case.
-    start_of_line: usize,
+    start_of_line: u64,
 }
 
 impl<I> LineColIterator<I>
@@ -32,15 +32,15 @@ where
         }
     }
 
-    pub fn line(&self) -> usize {
+    pub fn line(&self) -> u64 {
         self.line
     }
 
-    pub fn col(&self) -> usize {
+    pub fn col(&self) -> u64 {
         self.col
     }
 
-    pub fn byte_offset(&self) -> usize {
+    pub fn byte_offset(&self) -> u64 {
         self.start_of_line + self.col
     }
 }
